added some circuits + frontend modularisation
