+++ benchmarking-suite/moPro/mopro-example-app/flutter/lib/models/benchmark_item.dart
@@ -22,6 +22,10 @@ class BenchmarkResult {
   final Duration? provingTime;
   final Duration? verificationTime;
   final String? error;
+  
+  final int? proofSize;
+  final Map<String, dynamic>? memoryInfo;
+  final Map<String, dynamic>? batteryInfo;
 
   BenchmarkResult({
     required this.framework,
@@ -31,6 +35,9 @@ class BenchmarkResult {
     this.provingTime,
     this.verificationTime,
     this.error,
+    this.proofSize,
+    this.memoryInfo,
+    this.batteryInfo,
   });
 
   BenchmarkResult copyWith({
@@ -38,6 +45,9 @@ class BenchmarkResult {
     Duration? provingTime,
     Duration? verificationTime,
     String? error,
+    int? proofSize,
+    Map<String, dynamic>? memoryInfo,
+    Map<String, dynamic>? batteryInfo,
   }) {
     return BenchmarkResult(
       framework: framework,
@@ -47,6 +57,10 @@ class BenchmarkResult {
       provingTime: provingTime ?? this.provingTime,
       verificationTime: verificationTime ?? this.verificationTime,
       error: error ?? this.error,
+      proofSize: proofSize ?? this.proofSize,
+      memoryInfo: memoryInfo ?? this.memoryInfo,
+      batteryInfo: batteryInfo ?? this.batteryInfo,
     );
   }
 }
+
