+++ benchmarking-suite/moPro/mopro-example-app/flutter/lib/theme/app_theme.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class AppTheme {
+  static const Color primary = Color(0xFF5B56E6);
+  static const Color secondary = Color(0xFF1E40AF);
+  static const Color accent = Color(0xFF00BCD4);
+  static const Color danger = Color(0xFFFF6B6B);
+  static const Color warning = Color(0xFFFFA500);
+  static const Color success = Color(0xFF10B981);
+  static const Color background = Color(0xFFF5F7FA);
+  static const Color surface = Color(0xFFFFFFFF);
+  static const Color text = Color(0xFF1A1A1A);
+  static const Color textSecondary = Color(0xFF6B7280);
+  static const Color border = Color(0xFFE5E7EB);
+}
